requirements.txt
aioredis
amqp==5.1.1
asgiref==3.5.2
async-timeout
attrs==22.1.0
autobahn==22.7.1
Automat==20.2.0
backports.zoneinfo==0.2.1
billiard==3.6.4.0
cached-property==1.5.2
celery==5.2.7
certifi==2022.5.18.1
cffi==1.15.1
channels==3.0.5
channels-redis==3.3.1
charset-normalizer==2.0.12
click==8.1.3
click-didyoumean==0.3.0
click-plugins==1.1.1
click-repl==0.2.0
colorama==0.4.5
constantly==15.1.0
cryptography==37.0.4
daphne==3.0.2
Deprecated==1.2.13
Django==3.2.13
django-celery-beat==2.3.0
django-celery-results==2.4.0
django-cors-headers==3.13.0
django-filter==22.1
django-request-logging==0.7.1
django-timezone-field==5.0
django-widget-tweaks==1.4.8
djangorestframework==3.11.0
dnspython==2.2.1
docx2pdf==0.1.8
eventlet==0.33.1
greenlet==1.1.2
hiredis==2.0.0
hyperlink==21.0.0
idna==3.3
importlib-metadata==4.12.0
incremental==21.3.0
kombu==5.2.4
lxml==4.9.1
msgpack==1.0.4
packaging==21.3
Pillow==9.2.0
prompt-toolkit==3.0.29
psycopg2-binary==2.8.6
pyasn1==0.4.8
pyasn1-modules==0.2.8
pycparser==2.21
PyJWT==2.4.0
pyOpenSSL==22.0.0
pyparsing==3.0.9
python-crontab==2.6.0
python-dateutil==2.8.2
python-docx==0.8.11
python-dotenv==0.20.0
python-http-client==3.3.7
pytz==2022.1
redis==4.3.3
reportlab==3.6.11
requests==2.27.1
sendgrid==6.9.7
service-identity==21.1.0
six==1.16.0
sqlparse==0.3.1
starkbank-ecdsa==2.0.3
stripe==2.49.0
tqdm==4.64.1
Twisted==22.4.0
txaio==22.2.1
typing-extensions==4.2.0
tzdata==2022.2
urllib3==1.26.9
vine==5.0.0
wcwidth==0.2.5
wrapt==1.14.1
zipp==3.8.0
zope.interface==5.4.0
